wip: make package a poetry env and add gh action
