--- .travis.yml
@@ -1,43 +1,0 @@
-language: python
-
-os:
-  - linux
-
-python:
-  - "3.6"
-  - "3.7"
-
-env:
-  global:
-    - TOXENV=py
-
-notifications:
-  email: false
-
-before_install:
-  - mkdir -p ~/Downloads
-
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then curl -L https://julialang-s3.julialang.org/bin/linux/x64/1.3/julia-1.3-latest-linux-x86_64.tar.gz -o "$HOME/Downloads/julia.tar.gz"; fi
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then tar xzf "$HOME/Downloads/julia.tar.gz" -C "$HOME/Downloads"; fi
-  - if [ "$TRAVIS_OS_NAME" == "linux" ]; then cp -r "$(find "$HOME/Downloads" -maxdepth 2 -name "julia*" -type d | head -n 1)" "$HOME/julia"; fi
-
-  - if [ "$TRAVIS_OS_NAME" == "osx" ]; then curl -L https://julialang-s3.julialang.org/bin/mac/x64/1.3/julia-1.3-latest-mac64.dmg -o "$HOME/Downloads/julia.dmg"; fi
-  - if [ "$TRAVIS_OS_NAME" == "osx" ]; then hdiutil attach ~/Downloads/julia.dmg; fi
-  - if [ "$TRAVIS_OS_NAME" == "osx" ]; then cp -r /Volumes/Julia*/Julia*/Contents/Resources/julia $HOME/julia; fi
-  - if [ "$TRAVIS_OS_NAME" == "osx" ]; then hdiutil detach -force /Volumes/Julia*; fi
-
-  - rm -rf ~/Downloads/julia*
-  - export PATH="$HOME/julia/bin:$PATH"
-
-  - pip install --quiet tox
-
-script:
-
-  # Create an environment first so that we know what to use for $PYTHON:
-  - tox --notest
-
-  # PyCall requires to know the right Python executable to use:
-  - export PYTHON="$(echo $PWD/.tox/py*/bin/python)"
-  - echo "$PYTHON"
-
-  - tox -- --cov quickpomdps
