--- MANIFEST.in
@@ -1,4 +1,0 @@
-include README.md
-include LICENSE.md
-include docs/*.txt
-include quickpomdps/*.jl
