+++ pyproject.toml
@@ -1,0 +1,18 @@
+[tool.poetry]
+name = "quickpomdps"
+version = "0.0.1"
+description = "Describing and Solving POMDPs in Python"
+authors = ["rejuvyesh <mail@rejuvyesh.com>"]
+license = "MIT"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+julia = "^0.5"
+
+[tool.poetry.dev-dependencies]
+pytest = "^6.2.5"
+pytest-cov = "^2.12.1"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
