--- tox.ini
@@ -1,23 +1,0 @@
-[testenv]
-deps =
-    pytest
-    pytest-cov
-commands =
-    python -c 'import quickpomdps; quickpomdps.install()'
-    py.test \
-        --pyargs quickpomdps \
-        {posargs}
-whitelist_externals =
-    julia
-setenv =
-    # Do not use matplotlib GUI backend during tests.
-    MPLBACKEND = agg
-passenv =
-    # Allow a workaround for "error initializing LibGit2 module":
-    # https://github.com/JuliaLang/julia/issues/18693
-    SSL_CERT_FILE
-
-    # PyCall uses $PYTHON to configure Python executable.  It is
-    # useful to specify this in CI so that `diffeqpy.install()` can
-    # directly configure PyCall.
-    PYTHON
